Делает рефакторинг TasksController.php
