+++ services/TaskService.php
@@ -2,13 +2,16 @@
 
 namespace app\services;
 
+use app\models\CreateTaskForm;
 use app\models\Task;
 use app\models\User;
 use app\services\actions\ActionCancel;
 use app\services\actions\ActionFinish;
 use app\services\actions\ActionRefuse;
 use app\services\actions\ActionRespond;
+use Yii;
 use yii\base\Exception;
+use yii\db\StaleObjectException;
 use yii\helpers\Url;
 
 class TaskService
@@ -51,5 +54,79 @@ class TaskService
         }
 
         return $actionsMarkup;
+    }
+
+    /**
+     * @param CreateTaskForm $form
+     * @return Task
+     * @throws Exception
+     */
+    public function createTask(CreateTaskForm $form): Task
+    {
+        $locationService = new LocationService();
+
+        $task = new Task();
+        $task->loadDefaultValues();
+
+        $task->customer_id = Yii::$app->user->id;
+        $task->title = $form->title;
+        $task->description = $form->description;
+        $task->category_id = $form->category_id;
+        $task->budget = is_null($form->budget) ? null : (int) $form->budget;
+
+        if ($form->location !== '' && $locationService->isCityExistsInDB($form->city)) {
+            $task->city_id = $locationService->getCityIdByName($form->city);
+            $task->address = $form->address;
+            $task->coordinates = null;
+
+        }
+
+        $task->deadline_at = $form->deadline_at;
+        $task->save();
+
+        if ($form->location !== '') {
+            $locationService->setPointCoordinatesToTask($task->id, $form->lat, $form->long);
+        }
+
+        return $task;
+    }
+
+    /**
+     * @param Task $task
+     * @return void
+     * @throws StaleObjectException
+     * @throws \yii\db\Exception
+     */
+    public function refuseTask(Task $task): void
+    {
+        $userService = new UserService();
+        $executor = $task->executor;
+
+        $transaction = Task::getDb()->beginTransaction();
+
+        try {
+            $task->status = Task::STATUS_FAILED;
+            $task->update();
+            $executor->updateCounters(['failed_tasks_count' => 1]);
+            $executor->rating = $userService->countUserRating($executor);
+            $executor->status = User::STATUS_READY;
+            $executor->update();
+
+            $transaction->commit();
+        } catch (\Exception $e) {
+            $transaction->rollBack();
+            throw $e;
+        }
+    }
+
+    /**
+     * @param Task $task
+     * @return void
+     * @throws StaleObjectException
+     */
+    public function cancelTask(Task $task): void
+    {
+        $task->status = Task::STATUS_CANCELED;
+        $task->update();
     }
 }
